# openapi.json
{
  "components": {
    "schemas": {
...
      "AlertOut": {
        "description": "An official air-raid alert window (тривога -> відбій).",
        "properties": {
          "alert_type": {
            "title": "Alert Type",
            "type": "string"
          },
          "closed_reason": {
            "anyOf": [
              {
                "enum": [
                  "official",
                  "failsafe",
                  "dismissed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed Reason"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "scope": {
            "enum": [
              "city",
              "oblast"
            ],
            "title": "Scope",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "alert_type",
          "started_at",
          "provider"
        ],
        "title": "AlertOut",
        "type": "object"
      },
...
      "DismissedOut": {
        "description": "GET /admin/dismissed — recently admin-cancelled entities, for the\n'Повернути' (restore) list in the admin panel.",
        "properties": {
          "alerts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AlertOut"
            },
            "title": "Alerts",
            "type": "array"
          },
          "incidents": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IncidentOut"
            },
            "title": "Incidents",
            "type": "array"
          },
          "threats": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ThreatOut"
            },
            "title": "Threats",
            "type": "array"
          }
        },
        "title": "DismissedOut",
        "type": "object"
      },
...
      "EventDistrictIn": {
        "description": "PATCH /admin/events/{id} — admin fixes a mislocated sighting.",
        "properties": {
          "district_id": {
            "title": "District Id",
            "type": "integer"
          }
        },
        "required": [
          "district_id"
        ],
        "title": "EventDistrictIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentOut": {
        "description": "A coordinated attack with counts aggregated from its member threats.",
        "properties": {
          "alert_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Id"
          },
          "attack_types": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Attack Types",
            "type": "array"
          },
          "citywide": {
            "default": false,
            "title": "Citywide",
            "type": "boolean"
          },
          "classification": {
            "default": "unknown",
            "title": "Classification",
            "type": "string"
          },
          "decoy_suspected": {
            "default": false,
            "title": "Decoy Suspected",
            "type": "boolean"
          },
          "district_count": {
            "default": 0,
            "title": "District Count",
            "type": "integer"
          },
          "district_ids": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "District Ids",
            "type": "array"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "ended_reason": {
            "anyOf": [
              {
                "enum": [
                  "all_clear",
                  "alert_end",
                  "stale",
                  "dismissed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended Reason"
          },
          "has_hypersonic": {
            "default": false,
            "title": "Has Hypersonic",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "impact_count": {
            "default": 0,
            "title": "Impact Count",
            "type": "integer"
          },
          "notable": {
            "default": false,
            "title": "Notable",
            "type": "boolean"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_type": {
            "enum": [
              "shahed",
              "jet_drone",
              "missile",
              "ballistic",
              "unknown"
            ],
            "title": "Target Type",
            "type": "string"
          },
          "track_count": {
            "default": 0,
            "title": "Track Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "started_at",
          "target_type",
          "status"
        ],
        "title": "IncidentOut",
        "type": "object"
      },
...
      "ThreatOut": {
        "properties": {
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "closed_reason": {
            "anyOf": [
              {
                "enum": [
                  "destroyed",
                  "all_clear",
                  "stand_down",
                  "stale",
                  "dismissed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed Reason"
          },
          "confidence": {
            "default": 0.5,
            "title": "Confidence",
            "type": "number"
          },
          "corroboration_count": {
            "default": 1,
            "title": "Corroboration Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ThreatEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "has_conflict": {
            "default": false,
            "title": "Has Conflict",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "kind": {
            "default": "track",
            "enum": [
              "track",
              "impact"
            ],
            "title": "Kind",
            "type": "string"
          },
          "scope": {
            "default": "district",
            "enum": [
              "district",
              "city"
            ],
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "enum": [
              "unconfirmed",
              "tracking",
              "destroyed",
              "lost",
              "impact",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_count": {
            "default": 1,
            "title": "Target Count",
            "type": "integer"
          },
          "target_type": {
            "enum": [
              "shahed",
              "jet_drone",
              "missile",
              "ballistic",
              "unknown"
            ],
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "target_type",
          "status"
        ],
        "title": "ThreatOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kyiv Aerial Threat Tracker (MVP)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/alerts/{alert_id}/dismiss": {
      "post": {
        "operationId": "admin_dismiss_alert_admin_alerts__alert_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Dismiss Alert"
      }
    },
    "/admin/alerts/{alert_id}/restore": {
      "post": {
        "operationId": "admin_restore_alert_admin_alerts__alert_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Restore Alert"
      }
    },
...
    "/admin/dismissed": {
      "get": {
        "description": "Recently admin-cancelled threats / incidents / alerts, for the restore\n('Повернути') list in the admin panel.",
        "operationId": "admin_dismissed_admin_dismissed_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DismissedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Dismissed"
      }
    },
    "/admin/events/{event_id}": {
      "delete": {
        "description": "Remove a wrongly-attributed sighting. If it was the track's last event,\nthe now-empty track is dismissed; otherwise the parent incident's type is\nrecomputed from what remains.",
        "operationId": "admin_delete_event_admin_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Delete Event"
      },
      "patch": {
        "description": "Reassign a sighting to a different district (fix a mislocation). The\nthreat's vector/geometry are derived from its events at serialization, so a\nre-broadcast is all that's needed to update the map.",
        "operationId": "admin_move_event_admin_events__event_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDistrictIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Move Event"
      }
    },
...
    "/admin/incidents/{incident_id}/dismiss": {
      "post": {
        "operationId": "admin_dismiss_incident_admin_incidents__incident_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Dismiss Incident"
      }
    },
    "/admin/incidents/{incident_id}/restore": {
      "post": {
        "operationId": "admin_restore_incident_admin_incidents__incident_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Restore Incident"
      }
    },
...
    "/admin/threats/{threat_id}": {
      "patch": {
        "operationId": "admin_retype_threat_admin_threats__threat_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "threat_id",
            "required": true,
            "schema": {
              "title": "Threat Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreatTypeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Retype Threat"
      }
    },
    "/admin/threats/{threat_id}/dismiss": {
      "post": {
        "operationId": "admin_dismiss_threat_admin_threats__threat_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "threat_id",
            "required": true,
            "schema": {
              "title": "Threat Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Dismiss Threat"
      }
    },
    "/admin/threats/{threat_id}/restore": {
      "post": {
        "operationId": "admin_restore_threat_admin_threats__threat_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "threat_id",
            "required": true,
            "schema": {
              "title": "Threat Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Restore Threat"
      }
    },
...
  }
}